+++ pro/mcp/desktop/src/main/kotlin/org/http4k/mcp/internal/McpClientSecurity.kt
@@ -47,11 +47,11 @@ sealed interface McpClientSecurity : Security {
             oauthCredentials,
             tokenUri,
             backend,
-            ClientFilters.OAuthClientCredentials(oauthCredentials, scopes),
             Duration.ofSeconds(30),
             clock,
             ContentTypeJsonOrForm(),
             scopes,
+            ClientFilters.OAuthClientCredentials(oauthCredentials, scopes),
         )
     }